+++ packages/accounts-controller/src/utils.ts
@@ -39,8 +39,8 @@ export function keyringTypeToName(keyringType: string): string {
     case KeyringTypes.qr: {
       return 'QR';
     }
-    case KeyringTypes.cash: {
-      return 'Cash Account';
+    case KeyringTypes.money: {
+      return 'Money Account';
     }
     case KeyringTypes.snap: {
       return 'Snap Account';