+++ packages/cash-account-service/src/CashAccountService-method-action-types.ts
@@ -1,9 +1,0 @@
-import type { CashAccountService } from './CashAccountService';
-
-export type CashAccountServiceCreateCashAccountAction = {
-  type: `CashAccountService:createCashAccount`;
-  handler: CashAccountService['createCashAccount'];
-};
-
-export type CashAccountServiceMethodActions =
-  CashAccountServiceCreateCashAccountAction;