+++ packages/cash-account-service/src/index.ts
@@ -1,9 +1,0 @@
-export type {
-  CashAccountServiceActions,
-  CashAccountServiceMessenger,
-} from './types';
-export type {
-  CashAccountServiceCreateCashAccountAction,
-  CashAccountServiceMethodActions,
-} from './CashAccountService-method-action-types';
-export { CashAccountService } from './CashAccountService';