+++ packages/keyring-controller/CHANGELOG.md
@@ -9,8 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Add `KeyringTypes.cash` (`'Cash Keyring'`) to the `KeyringTypes` enum ([#8204](https://github.com/MetaMask/core/pull/8204))
-- Add `CashKeyring` (from `@metamask-previews/eth-cash-keyring`) as a built-in default keyring ([#8204](https://github.com/MetaMask/core/pull/8204))
+- Add `KeyringTypes.money` (`'Money Keyring'`) to the `KeyringTypes` enum ([#8204](https://github.com/MetaMask/core/pull/8204))
+- Add `MoneyKeyring` (from `@metamask/eth-money-keyring`) as a built-in default keyring ([#8204](https://github.com/MetaMask/core/pull/8204))
 
 ### Changed